Fixing type Error on Temperature Value (#12)

+++ src/Models/Request.swift
@@ -62,7 +62,7 @@ import MetaCodable
 	/// Higher values like 0.8 will make the output more random, while lower values like 0.2 will make it more focused and deterministic.
 	///
 	/// We generally recommend altering this or `top_p` but not both.
-	public var temperature: Int?
+	public var temperature: Double?
 
 	/// Configuration options for a text response from the model. Can be plain text or structured JSON data.
 	/// - [Text inputs and outputs](https://platform.openai.com/docs/guides/text)
@@ -125,7 +125,7 @@ import MetaCodable
 		reasoning: ReasoningConfig? = nil,
 		store: Bool? = nil,
 		stream: Bool? = nil,
-		temperature: Int? = nil,
+		temperature: Double? = nil,
 		text: TextConfig? = nil,
 		toolChoice: Tool.Choice? = nil,
 		tools: [Tool]? = nil,
